+++ packages/antwerp-ui/react-components/src/lib/molecules/autocomplete/Autocomplete.tsx
@@ -151,6 +151,7 @@ export function Autocomplete({
           name={name}
           label={label}
           onBlur={setValueBack}
+          onFocus={() => setIsOpen(true)}
           value={fieldValue}
           onChange={(e) => handleInput(e.target.value)}
           role="combobox"