+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Upgraded the `Icon` component so it can be used in SSR contexts
+- Upgraded the `Modal` component so it can be used in SSR contexts
+- Fixed an issue in the `Tooltip` component where it was possible to provide an invalid id attribute
+
 ## 7.12.0 - 2026-03-13
 
 ### Added